Update get_started.md
Added a bit on how to run on Windows. 

I'm not sure about this to be honest. The process is probably simpler than on MacOS (at least the last time I used MacOS), yet the description for MacOS is just "(although MacOS users will need to tell their system they really want to run it)". If you're expecting users to have that level of technical competence, then changing `MacOS` to `Windows/MacOS` would suffice, and the extra sentence isn't necessary.

Also I did this on the online editor without rulers, sorry if the line is too long.