"continuum hypothesis" -> "independence of the continuum hypothesis"
For me, "continuum hypothesis" could mean something as weak as "we stated the continuum hypothesis" (indeed "perfectoid spaces" above it means "we stated the definition of a perfectoid space"). Of all the questions one could possibly ask about the continuum hypothesis, independence is probably the hardest one; it's a highly nontrivial piece of mathematics which got Cohen a Fields Medal. I think we should stress that this is what was done.