Updates to Tags and Branches
@semorrison walked me through the process, so I went through the text to see if it is up to date, and made refinements where needed.

I removed the mentions of the latest-tag script, which I remove. Didn’t know it was so prominently advertised here.
Maybe if https://github.com/leanprover/lean4/pull/3137 works nicely we can report the latest tag this way.

Also mentioned `nightly-with-mathlib`, which I guess should be the default branch point for most contributers.