Clarify review process (#608)
I gave a shot at clarifying this text per the discussion in
[Zulip](https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/Steps.20after.20first.20approval.20in.20pull.20request.3F)

Some decisions I took:

- Replace the queue-redirect with a link to the queueboard (we might
want to update the #queue linkifier and/or this redirect?)
- Tried to avoid saying "reviewer" when it's not about someone in the
reviewers team.
- Tried to manage expectations with respect to the review process

Linking the maintainers/reviewers directly from this page might be
controversial, since it could lead to increases in direct messages to
them. I tried to mitigate this by adding the remark about the PR reviews
stream. I do think it is worthwhile being clear about who has what role,
because (as per the Zulip thread), this is part of what's causing
confusion.

---------

Co-authored-by: Yaël Dillies <yael.dillies@gmail.com>
Co-authored-by: grunweg <rothgang@math.uni-bonn.de>