feat: add a sidebar entry linking to the community guidelines (#627)
These ought to be augmented with a section on AI usage and spam (as was
posted on zulip); this can happen in a follow-up PR.

-------

Note: I am not sure if linking to a subsection of a page actually works
this way. Testing welcome (last time I tried to set up a local testing
environments, things failed miserably, so I cannot do so myself).