feat: add recent clarification to community guidelines (#628)
Clarify the community guidelines by including the recent annoucnement
[on
zulip](https://leanprover.zulipchat.com/#narrow/channel/113486-announce/topic/Community.20Policies/near/514574311)
in the written text. (This was always implicit policy, but it helps to
make it more explicit.)

Specify some further examples of unwelcome behaviour that is
disrespectful or disruptive, including using the community to complete
your coursework, use of sock puppets and AI abuse.
There may be further clarifications in the future, but the above is
already part of the maintainers' consensus.