update menu and installation instructions (#663)
* Remove `lake` from side menu (these are legacy instructions, made
obsolete by the installation guide. They are still accessible through
`get_started`)
* Add a direct link to `quickstart` in the sidebar
* Rename "set up Lean" to "More options". The idea of the "more options"
is that it also repeats the other two suggestions in the sidebar, and
gives a better description of the online options.
* Link to https://live.lean-lang.org/ on the "more options" page
* Prefer Codespaces over Gitpod on the "more options" page (I already
preferred it before, but now Gitpod is deprecated we definitely should
do this)
* Remove all the text about projects. That is now completely obsolete
with the quickstart guide, which will ask you to install a project like
Mathlib or Mathematics in Lean.

EDIT:
* On the main page, also directly link to `quickstart`
* On the main page, remove the link to Lean projects.