Add comments about what is in scope for mathlib (#526)
We add some comments about various reasons why a Lean development might
or might not be a good fit for mathlib. The tl;dr: to be a good fit for
mathlib it should be the kind of material which is likely to be studied
in mathematics departments, and ideally at least one maintainer should
have a good knowledge of the area.

---------

Co-authored-by: Kim Morrison <477956+kim-em@users.noreply.github.com>
Co-authored-by: Floris van Doorn <fpvdoorn@gmail.com>
Co-authored-by: Bryan Gin-ge Chen <bryangingechen@gmail.com>