feat: add workflow to build site on PRs (#702)
This is a version of #571 which puts the `pull_request` trigger in a
separate workflow without access to secrets. (The build works fine
without them, except for getting Zulip users to put on the map, and that
just throws a harmless warning.)

I also pinned the used workflow versions and added steps to upload the
build files as an artifact. (Later on we might switch from deploying by
pushing to `master` to deploying via artifact with
actions/deploy-pages.)