feat: documentation overview page (#731)
This page is intended to be a complete collection of all the pieces of
documentation for Mathlib users and Lean users more generally. I have
gone through everything findable on the community website, and quite a
few suggestions on the [Zulip
chat](https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/Mathlib.3A.20the.20Missing.20Manuals)
and tagged and categorized them all.

On the rendered page, we can see a diagram of the [Divio classification
system](https://docs.divio.com/documentation-system/). I also added the
possibility to filter the entries by tags.

Here is the rendered page, which works standalone in your browser (only
missing a bit of its formatting):

[documentation.html](https://github.com/user-attachments/files/23459583/documentation.html).