feat(documentation.html): collapse the documentation quadrants
This implements a suggestion by Chris Henson on Zulip: https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/Discussion.3A.20Mathlib.20documentation.20overview/near/556896150

Making each category collapsed means that there is a lot less to scroll through before reaching the right category. It also means we can put the category description with the category being referred to.

The categories automatically open when you select a tag, otherwise it would be weird that you click on a tag but nothing seems to happen.

I would have liked to make the headers have the collapse/expand arrow, instead of the description, but it seems like starting a `<summary>` with a `<h2>` is not really supported, since the arrow goes on its own line before the header.