fix: update sample module doc-string for the introduction of the module system
The documentation style guide quotes a full module header as an example:
update that to the current state in mathlib, by using the module system instead.

One may argue that this is too early, given the module system is still
experimental. On the other hand, that restriction will be lifted soon,
and mathlib *is* using the module system already and not going back.